New system contract first revision

+++ libraries/chain/apply_context.cpp
@@ -18,7 +18,7 @@ void apply_context::exec_one()
       auto native = mutable_controller.find_apply_handler(receiver, act.account, act.name);
       if (native) {
          (*native)(*this);
-      } 
+      }
       else if (a.code.size() > 0) {
          try {
             mutable_controller.get_wasm_interface().apply(a.code_version, a.code, *this);
@@ -181,14 +181,14 @@ void apply_context::require_recipient( account_name code ) {
 
 
 /**
- *  This will execute an action after checking the authorization. Inline transactions are 
+ *  This will execute an action after checking the authorization. Inline transactions are
  *  implicitly authorized by the current receiver (running code). This method has significant
  *  security considerations and several options have been considered:
  *
  *  1. priviledged accounts (those marked as such by block producers) can authorize any action
  *  2. all other actions are only authorized by 'receiver' which means the following:
  *         a. the user must set permissions on their account to allow the 'receiver' to act on their behalf
- *  
+ *
  *  Discarded Implemenation:  at one point we allowed any account that authorized the current transaction
  *   to implicitly authorize an inline transaction. This approach would allow privelege escalation and
  *   make it unsafe for users to interact with certain contracts.  We opted instead to have applications
@@ -196,9 +196,9 @@ void apply_context::require_recipient( account_name code ) {
  *   can better understand the security risk.
  */
 void apply_context::execute_inline( action&& a ) {
-   if ( !privileged ) { 
+   if ( !privileged ) {
       if( a.account != receiver ) {
-         controller.check_authorization({a}, flat_set<public_key_type>(), false, {receiver}); 
+         controller.check_authorization({a}, flat_set<public_key_type>(), false, {receiver});
       }
    }
    _inline_actions.emplace_back( move(a) );
@@ -233,7 +233,7 @@ void apply_context::execute_deferred( deferred_transaction&& trx ) {
                break;
             }
          }
-         if( check_auth ) 
+         if( check_auth )
             controller.check_authorization(trx.actions, flat_set<public_key_type>(), false, {receiver});
       }
 
@@ -245,7 +245,7 @@ void apply_context::execute_deferred( deferred_transaction&& trx ) {
    } FC_CAPTURE_AND_RETHROW((trx));
 }
 
-void apply_context::cancel_deferred( uint32_t sender_id ) {
+void apply_context::cancel_deferred( uint64_t sender_id ) {
    results.canceled_deferred.emplace_back(receiver, sender_id);
 }
